requirements.txt
PyQt5==5.15.9
requests==2.31.0
websocket-client==1.6.4
pandas==2.0.3
pytest==7.4.3
pytest-cov==4.1.0